lxd: Move replicator run logic to separate file.
